// API_Client/client/models/payment-stats-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PaymentStatusBreakdownItemDto } from './payment-status-breakdown-item-dto';

export interface PaymentStatsDto {
    'breakdown': Array<PaymentStatusBreakdownItemDto>;
    /**
     * პერიოდში შექმნილი გადახდების საერთო რაოდენობა
     */
    'total': number;
    /**
     * COMPLETED სტატუსის წილი პროცენტებში ყველა გადახდას შორის — null, თუ პერიოდში გადახდები არ ყოფილა
     */
    'successRatePercent': number | null;
}