// API_Client/client/models/payment-status-breakdown-item-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface PaymentStatusBreakdownItemDto {
    'status': PaymentStatusBreakdownItemDtoStatusEnum;
    'count': number;
}

export const PaymentStatusBreakdownItemDtoStatusEnum = {
    Created: 'created',
    Processing: 'processing',
    Completed: 'completed',
    Rejected: 'rejected',
    Refunded: 'refunded',
    PartialCompleted: 'partial_completed',
} as const;

export type PaymentStatusBreakdownItemDtoStatusEnum = typeof PaymentStatusBreakdownItemDtoStatusEnum[keyof typeof PaymentStatusBreakdownItemDtoStatusEnum];